Add files via upload
The script requires a serialized image input from a Scanning Electron Microscope with Backscattered Electron detector (SEM-BSE) that have been corrected for shading (16-bit greyscale). The image input for doing the pre-calibration for shading is often found in the raw image tiles of a whole-slide SEM-BSE. In the case of the TIMA software the usual path is 'Project source folder' > 'Name.timeproj.Datasets > 'tokenized names' (pick the one) > 'fields'. Use 'switch names_v1.m' script to parse those individual image tiles within the sub-folders. 
The technical explanation and purpose of recolouring a SEM-BSE into RGB (non-linearly) is given in the paper (Acevedo Zamora and Kamber, 2023). Please, cite if using this code. Thanks.