Add files via upload
Scripts for interpreting labeled images exported from QuPath after the application of the 'Pixel Classifier'. The initial name of the mineral phases used for segmentation (class names) can be updated to a more adequate naming (for publication) and the thin section fringes can be cropped out (manual editing) to unbias the modal mineralogy estimates.